requirements.txt
djangorestframework
knox
django